trace_test.go => internal/cmd/gtrace/trace_hooks_test.go
